service/bot/requirements.txt
anyio==4.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
isort==5.13.2
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
platformdirs==4.1.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tzdata==2023.4
urllib3==2.2.1
